+++ CHANGELOG.md
@@ -108,5 +108,6 @@
 
 ### 1.1.9
 
+- Fixed the Potion of Cell Growth recipe setting the entity when it shouldn't (#14)
 - Support Slimes now have a 100% chance to give the Slimy Death Gene. The file previously pointed at the entity `minecraft:slime` rather than `geneticsresequenced:support_slime`
 - Moved to better practices (using less lazy values, mostly)